+++ packages/plugin-sdk-value/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 7.2.1
+
+### Patch Changes
+
+- Updated dependencies [[`b18469f`](https://github.com/portabletext/editor/commit/b18469f7c73ef036c633142f3864ed16a70fa93c)]:
+  - @portabletext/editor@7.10.16
+
 ## 7.2.0
 
 ### Minor Changes